cleaned up the code for A E S T H E T I C S
