+++ package.json
@@ -4,6 +4,7 @@
   "description": "A rebuilt version of lolfy",
   "main": "index.js",
   "scripts": {
+    "test": "jest",
     "dev": "webpack-dev-server",
     "build": "webpack",
     "server": "http-server ./src/client/dist",
@@ -29,6 +30,7 @@
   "homepage": "https://github.com/zzmarkzz321/LoLFYv2#readme",
   "devDependencies": {
     "babel-core": "^6.26.0",
+    "babel-jest": "^22.0.4",
     "babel-loader": "^7.1.2",
     "babel-plugin-transform-runtime": "^6.23.0",
     "babel-polyfill": "^6.26.0",
@@ -40,6 +42,7 @@
     "html-webpack-plugin": "^2.30.1",
     "http-server": "^0.10.0",
     "path": "^0.12.7",
+    "react-test-renderer": "^16.2.0",
     "webpack": "^3.10.0",
     "webpack-dev-server": "^2.9.7"
   },
