+++ src/client/components/App.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React from 'react';
 import { HashRouter, Route, Switch } from 'react-router-dom';
 
 import { Home } from "./Home";
@@ -9,8 +9,7 @@ export const App = () => (
     <HashRouter>
         <Switch>
             <Route path="/" exact component={Home}/>
-            {/*Remove in production*/}
-            <Route name="results" path="/results" component={Results}/>
+            <Route path="/results" component={Results}/>
             <Route component={NotFound}/>
         </Switch>
     </HashRouter>
