+++ src/client/tests/Home.test.js
@@ -1,0 +1,3 @@
+import React from 'react';
+import { Home } from "../components/Home/Home";
+import renderer from 'react-test-renderer';
