+++ src/server/tests/getSummonerInfo.test.js
@@ -1,0 +1,9 @@
+'use strict';
+
+const matchInfoMock = require('../mockData/matchInfo');
+const recentMatchesMock = require('../mockData/recentMatches');
+const summonerInfoMock = require('../mockData/summonerInfo');
+
+test('_formatMatchData formats data correctly', done => {
+    console.log(JSON.stringify(summonerInfoMock, null, 2));
+});
